optimize the case where we're writing a single value in LevelDB
Benches don't show a significant difference in performance with this patch, neither for the single-value case nor for the multiple-value cases.  Nevertheless, reviewers of the code and its results will expect this optimization, given that Writebatch could be more expensive than writing directly to the database in theory.

Even better would be to write a separate function that doesn't use Writebatch and compare it to one that does.  Perhaps I'll do that later.